requirements.txt
pyTelegramBotAPI
google-generativeai
flask